IO layout cleanup
Changed location of bal/unbal buffers from selector board to daughter board. Generalized IO PCB to accept daughter boards, not specific to connector type or bal/unbal. The idea is that as many unbalanced RCAs can be populated on the main IO board as desired and jumper cables can also be added to daughter boards as needed for different kinds of IO, but the interface to the selector board does not need to change to allow this.